LayoutTests/http/wpt/wk-web-animations/timing-model/document-timeline.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check that the DocumentTimeline interface is defined and that the Document has one by default</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
    assert_own_property(window, "DocumentTimeline");
}, "The DocumentTimeline interface is defined.");

test(t => {
    assert_inherits(document, "timeline");
}, "The Document interface exposes a timeline property.");

test(t => {
    assert_true(document.timeline instanceof DocumentTimeline);
}, "The document timeline is a DocumentTimeline.");

test(t => {
    assert_true(document.timeline instanceof AnimationTimeline);
}, "The document timeline inherits from AnimationTimeline.");

test(t => {
    const timeline = document.timeline;
    assert_equals(timeline, document.timeline);
}, "The document timeline is a single object.");

</script>
</body>
